implemented ValidatValueAgainstDataTypeSpec
